requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2022.9.24
charset-normalizer==2.1.1
flake8==5.0.4
idna==3.4
importlib-metadata==4.2.0
mccabe==0.7.0
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==0.21.0
python-telegram-bot==13.14
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
tornado==6.1
typing-extensions==4.4.0
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
zipp==3.10.0
